+++ join.cpp
@@ -12,7 +12,7 @@
 //#define H1(X) (X & 0xFF)
 //#define H2(A, B) ((A) / (B))
 
-int32_t twoInLSB;	// 2^HASH_LSB
+int64_t twoInLSB;	// 2^HASH_LSB
 
 /* First parsing through table
  * Find hash position and increase histogram value by one
@@ -25,21 +25,21 @@ int32_t twoInLSB;	// 2^HASH_LSB
 
 void hash_relation(relation_info* newRel, relation *rel){
 
-	newRel->histogram = (int32_t*)calloc(twoInLSB, sizeof(int32_t));
-	for (int32_t i = 0; i < rel->num_tuples; i++){
-		int32_t position = rel->tuples[i].payload & (twoInLSB-1);
+	newRel->histogram = (int64_t*)calloc(twoInLSB, sizeof(int64_t));
+	for (int64_t i = 0; i < rel->num_tuples; i++){
+		int64_t position = rel->tuples[i].payload & (twoInLSB-1);
 		newRel->histogram[position]++;
 	}
 
-	int32_t *sumHistogram = (int32_t*)malloc(twoInLSB * sizeof(int32_t));
+	int64_t *sumHistogram = (int64_t*)malloc(twoInLSB * sizeof(int64_t));
 	sumHistogram[0] = 0;
 	for (int i = 1; i < twoInLSB; i++)
 		sumHistogram[i] = sumHistogram[i-1] + newRel->histogram[i-1];
 
 	newRel->tups.num_tuples = rel->num_tuples;
 	newRel->tups.tuples = (tuple*)malloc((rel->num_tuples)*sizeof(tuple));
-	for (int32_t i = 0; i < rel->num_tuples; i++){
-		int32_t position = rel->tuples[i].payload & (twoInLSB-1);
+	for (int64_t i = 0; i < rel->num_tuples; i++){
+		int64_t position = rel->tuples[i].payload & (twoInLSB-1);
 		memcpy(newRel->tups.tuples+sumHistogram[position], &rel->tuples[i], sizeof(tuple));
 		sumHistogram[position]++;
 	}
@@ -56,24 +56,24 @@ void hash_relation(relation_info* newRel, relation *rel){
 void join_buckets(result* list, relation_info* small, relation_info* big, int begSmall, int begBig, int bucketNo, int orderFlag) {
 	int hashValue = next_prime(small->histogram[bucketNo]);
 
-	int32_t* bucket = (int32_t*)malloc(hashValue * sizeof(int32_t));
+	int64_t* bucket = (int64_t*)malloc(hashValue * sizeof(int64_t));
 	for (int i = 0; i < hashValue; i++)
 		bucket[i] = -1;
 
-	int32_t* chain = (int32_t*)malloc(small->histogram[bucketNo] * sizeof(int32_t));
+	int64_t* chain = (int64_t*)malloc(small->histogram[bucketNo] * sizeof(int64_t));
 	int endSmall = begSmall + small->histogram[bucketNo];
 
-	for (int32_t i = begSmall; i < endSmall; i++){
-		int32_t position = (small->tups.tuples[i].payload) % hashValue;
+	for (int64_t i = begSmall; i < endSmall; i++){
+		int64_t position = (small->tups.tuples[i].payload) % hashValue;
 		int previousValue = bucket[position];
 		bucket[position] = i-begSmall;
 		chain[i-begSmall] = previousValue;
 	}
 	int endBig = begBig + big->histogram[bucketNo];
 
-	for (int32_t i = begBig; i < endBig; i++) {
+	for (int64_t i = begBig; i < endBig; i++) {
 		int bigHash = (big->tups.tuples[i].payload) % hashValue;
-		int32_t position = bucket[bigHash];
+		int64_t position = bucket[bigHash];
 		while (position != -1) {
 			if (big->tups.tuples[i].payload == small->tups.tuples[position+begSmall].payload) {
 				if (orderFlag)
