+++ result_list/resultList.h
@@ -2,8 +2,8 @@
 #define _RESULT_LIST_
 
 struct key_tuple{
-	int32_t keyR;
-	int32_t keyS;
+	int64_t keyR;
+	int64_t keyS;
 };
 
 struct bucket_info {
@@ -12,14 +12,14 @@ struct bucket_info {
 };
 
 struct result{
-	int32_t capacity;			// Capacity of each & every bucket_info
-	int32_t size;				// Size of current bucket_info (all the other bucket_infos and memory behind are full)
+	int64_t capacity;			// Capacity of each & every bucket_info
+	int64_t size;				// Size of current bucket_info (all the other bucket_infos and memory behind are full)
 	bucket_info* head;			// Current bucket_info
 };
 
 void init_list(result*);
 void empty_list(result*);
-void add_result(result*, int32_t, int32_t);
+void add_result(result*, int64_t, int64_t);
 void print_list(result*);
 
 #endif
